requirements.txt
alembic==1.17.1
amqp==5.3.1
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asyncpg==0.30.0
bcrypt==5.0.0
billiard==4.2.2
cachetools==6.2.1
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.121.0
fastapi-limiter==0.1.6
google-auth==2.42.1
google-genai==1.48.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jose==1.0.0
kombu==5.5.4
Mako==1.3.10
MarkupSafe==3.0.3
packaging==25.0
passlib==1.7.4
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
redis==7.0.1
requests==2.32.5
resend==2.19.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.3
stripe==13.2.0
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
vine==5.1.0
wcwidth==0.2.14
websockets==15.0.1
